Rename create/edit alert dialog to form
