+++ web-common/src/components/alert-dialog/alert-dialog-cancel.svelte
@@ -4,7 +4,7 @@
   import { cn } from "@rilldata/web-common/lib/shadcn";
 
   type $$Props = AlertDialogPrimitive.CancelProps;
-  type $$Events = AlertDialogPrimitive.CancelEvents;
+  // type $$Events = AlertDialogPrimitive.CancelEvents;
 
   let className: $$Props["class"] = undefined;
   export { className as class };
